+++ blender_manifest.toml
@@ -2,11 +2,11 @@ schema_version = "1.0.0"
 
 # Manifest file for MeshLint following documentation at, 
 # https://docs.blender.org/manual/en/latest/advanced/extensions/getting_started.html
-# The id is case sensitive with directory name!
+# The id is case-sensitive with directory name!
 
 id = "MeshLint"
 name = "MeshLint"
-version = "0.1.1"
+version = "0.1.2"
 tagline = "MeshLint is like spell-checking for your Meshes"
 # "description": "Check objects for: Tris / Ngons / Nonmanifoldness / etc"
 # "location": "Object Data properties > MeshLint"
@@ -29,7 +29,9 @@ copyright = ["2012 rking"]
 paths_exclude_pattern = [
   "__pycache__/",
   "testblends",
+  "tests",
   "/.*/",
   "/.*",
   "/*.zip",
+  "/*.blend*",
 ]
